Conversion for mime types other than application/json (the 1st registered) was broken in legacy Gecko (pre-2.0).
This is a proper fix for the factory exists error in stream converter registration.
The NS_ERROR_FACTORY_EXISTS exception is only thrown in newer Gecko for some reason (only just noticed), the workaround of setting the factory to null actually broke registration of more than 1 stream converter in older Gecko.
Also, shouldn't be using for..in on an array.