Improve API of prefs module.
Can now listen to changed preferences on a whole branch, not just single keys.
Getting a preference that doesn't exist will now return null rather than throw an exception.
A list of all preferences in a branch can be retrieved as an array.
A preference can be unset.
And finally, old exception handlers full of fail...