Add support in prefs module for listening to multiple preferences below the branch root.
Previously, listening for "" has triggered for all preference changes in the branch and any other string triggered for preference changes exactly matching, now we can end any string with a dot and this will trigger for all preferences that start with the string. This will allow us to listen for changes in sub-branches whilst continuing to use the single, efficient preference observer on the root branch.